Move budgeting test code into v4.
